Move repositories to service folder
